Fabric port for Minecraft 1.21.1
Full port of Epic Fight mod from NeoForge to Fabric 1.21.1.

Build system:
- Migrate from NeoGradle (build.gradle.kts) to Fabric Loom (build.gradle)
- Fabric Loader 0.19.3, Fabric API 0.116.15+1.21.1, Java 21
- Replace access transformers with access wideners
- Replace gradle/build-logic convention plugins with flat Loom config

Platform layer:
- Replace NeoForge DeferredRegister with Fabric deferred registry shims
  (DeferredRegisterShim, DeferredHolderShim, RegistryCallback)
- Replace NeoForge event bus with Fabric API callbacks and mixins
- Replace ForgeConfigAPIPort with native NightConfig-backed config (ModConfigSpec)
- Replace NeoForge network channels with Fabric Networking API payloads
- Replace @OnlyIn/Dist with Fabric @Environment/EnvType
- Replace NeoForge capability system with duck interfaces (EntityExtension,
  BlockStateExtension) and mixins
- Replace neoforge.mods.toml with fabric.mod.json
- Add platform abstraction (FabricModPlatform, FabricClientModPlatform)

Mod compatibility:
- Add Trinkets integration (Fabric equivalent of Curios)
- Add Controlify compile stubs for optional controller support
- Add AzureLib armor renderer compile stubs
- Remove NeoForge-only compat: KubeJS, Vampirism, Werewolves, Curios
- PlayerRevive: waiting for CurseForge approval (playerrevive-refabric)
- Retain cross-platform compat: GeckoLib, playerAnimator, Female Gender,
  Iris, Sodium, Better Third Person, First-person Model, Shoulder Surfing,
  3D Skin Layers, JEI

Rendering:
- Port SeparateTransforms baked model via Fabric model loading plugin
- Add EpicFightFrustumHolder for client-side frustum access
- Add mixin accessors for client internals (KeyMapping, LightTexture,
  ElytraLayer, ItemRenderer, etc.)

Initialization:
- Fail fast on required registration (networking, config, registries,
  entity patches, attributes, spawn placements)
- Catch only optional integrations (MCreator scan, EpicSkins, compute
  shaders, Controlify)

Bug fixes:
- Fix flat shading on held items and misread baked lightmap UV
- Fix left click never breaking blocks in Epic Fight mode
- Fix inverted block outline cancellation
- Fix config file collision with ForgeConfigAPIPort
- Move resource packs from packs/ to resourcepacks/ for Fabric convention
- Move emote data files out of nested epicfight/epicfight/ directory

+++ src/main/java/yesman/epicfight/client/gui/BattleModeGui.java
@@ -118,7 +118,7 @@ public class BattleModeGui {
 			String skillName = Component.translatable(String.format("skill.%s.%s", rl.getNamespace(), rl.getPath())).getString();
 			
 			int stringWidth = this.minecraft.font.width(skillName);
-			guiGraphics.drawString(this.minecraft.font, skillName, (pos.x + 120 - stringWidth * 0.5F), pos.y - 12, 16777215, true);
+			guiGraphics.drawString(this.minecraft.font, skillName, (int)(pos.x + 120 - stringWidth * 0.5F), pos.y - 12, 16777215, true);
 		}
 	}
 	
@@ -137,9 +137,9 @@ public class BattleModeGui {
 			if (slot == SkillSlots.WEAPON_INNATE) {
 				continue;
 			}
-			
+
 			SkillContainer container = playerpatch.getSkill(slot);
-			
+
 			if (!container.isEmpty()) {
 				if (!this.skillIcons.contains(container) && container.getSkill().shouldDraw(container)) {
 					this.skillIcons.add(container);
@@ -195,7 +195,7 @@ public class BattleModeGui {
 		}
 		
 		SkillContainer container = playerpatch.getSkill(SkillSlots.WEAPON_INNATE);
-		
+
 		if (!container.isEmpty() && container.getSkill().shouldDraw(container)) {
 			float partialTick = deltaTracker.getGameTimeDeltaPartialTick(false);
 			Vec2i pos = ClientConfig.getWeaponInnatePosition();
